Reorganize data to be independent of web server elements
In preparation for developing and testing the scheduling algorithm,
make a clean separation between the bits that are necessary for the
scheduling algorithm itself, and those which are used to serve
webpages.

At the moment, this includes users and discussions. Sessions are
web-focused and will not be contained here.

Move the "saving" functionality into the create / update code
themselves, rather than calling it in the web-handling code.

Also get rid of pointless interfaces for the store; this is almost
certainly an artifact of the book trying to use an example interface.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>