Refactor view URLs, add user view
Try to avoid a multiplication of functions by having discussion and
user uids, and view and edit functionalities of each, go through a
single function, HandleUid.  Handle just "view" for now.

Make discussion/notfound into a generic uid/notfound (but pass it the
type if thing that wasn't found so it displays differently).

Implement a user/view template.  Pass a .Display item into the */view templates.

Signed-off-by: George Dunlap <george.dunlap@citrix.com.