Allow users to express an interest in discussions
Break down "discussion/item" template into "discussion/item-short" for
lists, and "discussion/item-full".  Only allow interest to be set from
"full" pages.

Rather than simple "Attend" button, give four pre-made interest
levels, mapping to 0, 25, 75, and 100.  But only display this if there
is a user logged in, and they are not the administrator.

To make this possible, add "IsUser", "IsAdmin", and "Interest"
elements to DiscussionDisplay.

Add HandleUidPost to handle POST methods relating to UIDs.  This
unfortunately has to be handled by the main renderer, or it will come
up with "Invalid Method".

Other minor changes:
- Rename "IsMe" to "IsMine" for consistency
- Move 'Owner' to top, make text muted
- Use "user/link" template for schedule as well.  This requires adding
discussion ID to DiscussionScheduleDisplay.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>