Add basic Locations information
This isn't very clean, but we just need something that works.

Allow Locations to be either a "place" or not; places hold only a
single discussion, non-places hold an arbitrary amount.

Make LocationStore a slice so that we have an ordered preference in
which to put things.  At the moment, we add locations in the order we
want; there must be only a single !IsPlace which is at the very end.
If we extend this we should insert new locations at the right point in
the array.

Also, rather than trying to duplicate a "TimetableDisplay" structure
(which would be almost entirely identical to the Timetable structure),
just copy the whole Location structure into TimetableDiscussion.  (We
at least need the ID so that the template can make a link to the
location.)

Place the sessions in order by number of attendees.

Add location links to the schedule; handling location uids will be
done later.

While here, add labels to the attendee / interest badges.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>