Add timetable
Separate concept of a "schedule" (which is an abstract series of slots)
from a "timetable", which is a concrete mapping of specific timeslots that
happen over specific days (and may contain breaks with no sessions).

Hard-code timeslots (and timeslots) to current XenSummit schedule.

The "leaf" of a timetable includes information about 

Modify schedule/view template to print timetable instead of schedule.

Get rid of schedule display code and structures from schedule.go; and call
TemplatePlace(sched) after completing a schedule.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>