Limit number of discussions, as well as duplicate titles
Add a DiscussionStore.Iterate() method that calls a callback function
on each discussion, stopping when it returns an error (and returning
that error).

Use this to check for duplicate titles, as well as making sure that nobody
has more sessions than there are schedule slots.  (This is a sensible limit, as
we assume each person will want to attend their own sessions.)

Add errors for these two conditions as well.

Finally, teach NewTestDiscussion to handle the new failures
gracefully.

While we're here, mention the verification code in README.md

Signed-off-by: George Dunlap <george.dunlap@citrix.com>