Allow session deletion
Add DeleteDiscussion() function, which will remove an event from the
schedule and update the timetable (if a schedule exists), remove it
from user interest, and finally delete the discussion.

Add a "Delete" button next to the "Edit" button for those with MayEdit
privelege on a discussion.  Add a 'confirm' page.

Implement a UserStore.Iterate() in order to iterate over all users
looking for interest to remove.

Change the (previously unused) Delete function to take a DiscussionID
instead.

While here:
 - Give the 'Edit' button a color
 - Delete vestigal elements of some structs (`Attending` from
   TimetableDiscussion and `slot` from TimetableSlot
 - Implement a couple of other Users. methods in terms of Iterate.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>