Allow session slot restriction
Only visible / modifiable by the admin.

Add PossibleSlots []bool to the Discussion, initialized to
Event.ScheduleSlots length, all set to true.

Add IsAdmin to the DiscussionDisplay, along with an array of
DisplaySlots (which give the day/time and current status of each
slot).  Add Timetable.FillPossibleSlots() to convert []bool into
[]DisplaySlots with times and so on.

Display possible slots in "discussion/view" if IsAdmin.

Pass DiscussionDisplay to "edit" actions as well as "view".  Add
"DescriptionRaw" so that we can edit the raw description there.

Display possible slots in "discussion/edit" as a list of labelled
tickboxes.

Parse these in HandleUidPost itype == "discussion" action == "edit"
into a []bool, and pass to UpdateDiscussion (which has been extended
to accept this argument).

Randomly disable discussion slots in NewTestDiscussion() for testing.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>