Make a distinction between "user data" and "event data" when resetting
"Event data" is the schedule, timetime, and location -- currently either generated or hard-coded.

"User data" is data entered by the users.

Add a button to reset 'event data' on the main Admin console.

The "Clear User Data" still clears all the data, because the schedule
/ timetable doesn't make much sense without it.

While here, fix the layout for the admin/testing page.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>