Show discussions owned by a user
Add DiscussionStore.GetListUser(u, cur), which return a
[]*DiscussionDisplay tweaked to display to user 'cur', owned by user
'u'.

Add a list to UserDisplay, and an option to User.GetDisplay() to put
together such a list, and set it to 'true' when viewing an individual
user. (This must be called `List` so that it can be passed to the
discussion/list template.)

Add the "discussion/list" template at the bottom of the "user/view"
template.

While here, make the user 'edit' link into a proper button.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>