Allow admins to change discussion owners
Add DiscussionDisplay.AllUsers.  Fill it in when the current user is
an admin.

Modify template discussion/form, so that it presents a list of users
to select from to change.  For ease of implementaiton, add current
user (selected) on the top.  This means the current user will show up
in the list twice, but it's not really worth trying to work around.

Process that value in HandleUidPost() only if the current user is an
admin; otherwise set it to the current user.

Add "ownerID" as a field to UpdateDiscussion.  If the ID is changing,
then set the ID and also set the new owner's interest to 100.

This fixes a potential bug where an unauthorized user may be able to
modify the PossibleSlots schedule.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>