Display time and location in long discussion view if available
Add lower-cased location and time to Discussion as a "cache" for
displaying.  Run Timetable.Place() on load to fill this information
in.

"Time" in this case will be both the day and time.  Pass the day name
in to TimetableSlot.PlaceSlot() to make this possible.

Expose this information, if present, in "discussion/item-full".

Signed-off-by: George Dunlap <george.dunlap@citrix.com>