Allow 'locking' sessions so that they won't be rescheduled
Refactor "DisplaySlot" to be more generic:
- 'Checked' rather than 'Possible'
- FillDisplaySlots to create

Add a LockedSlots []boo to EventStore and initialize on creating and
reset.

Pass a []DisplaySlots of this to the admin console template.

Add a form indicating which slots are locked, and a button to submit
changes.

Factor out the loop for checking into FormCheckToBool.

Add a 'setLocked' admin handler to make updates to LockedSlots.

When scheduling, if a slot is locked:
- Copy the old one verbatim into the new one
- Exclude all the discussions in those slots from consideration
- Skip those slots when placing discussions

Signed-off-by: George Dunlap <george.dunlap@citrix.com>