Report locked days as Final, others as Tentative
Put LockedSlots into its own type.

Make a Set() method on the type so that we can do triggers when it's
updated, and use that to set it in HandleAdminAction().

Add TimetableDay.IsFinal. Add Timetable.UpdateIsFinal() to go through
and set this when all the slots in that day are locked.  Call it from LockedSlots.Set().

In the schedule template, put 'Final' or 'Tentative' badges after the day name.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>