Show tentative / final badge on the discussion view
Add an up-pointer from TimetableSlot to TimetableDay; make it
non-serialized, and so initialize it when placing and initializing.

Replace the "pre-cooked" Discussion.time with a pointer to the slot.

When creating DiscussionDisplay, use that backpointer to generate the
day and time, as well as determine if the day is finalized or not.

Factor out the code which displays final / tentative badge, and call
it to the discussion view page, with the new IsFinal element.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>