Reduce overhead of logging when not in use, disable validation
Using profiling, it turned out that a reasonably large amount of the
time was spent in log functions, or scoring to print out information
from log functions.

Get red of SchedDebugVerbose entirely, and wrap all calls with checks
for OptSchedDebugVerbose.  Also, on hotpaths for SchedDebug, check for
OptSchedDebug.

Also disable validation by default.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>