Make a local snapshot of key information for scheduling
In preparation for running scheduling asynchronously, take a snapshot
of all necessary information to work without holding the mutex.

To do this:

Add "up" pointers from slots to the main schedule.  Restore these on
load, and set them appropriately when generating or copying schedules.

Create SearchStore to store the key information needed for the search:
- A deep copy of Users, Discussions, and LockedSlots
- A reference to the schedule current at the time of the snapshot
(from which to copy locked slots)
- A list of discussions not in locked slots, which need to be
scheduled

Use encoding/gob to efficiently do deep copies of Users and
Discussions.

Add a 'store' element to the Schedule.  This should only be set while
the schedule is being generated.  Add a function to return where
callees should look for user and discussion stores (either the
snapshot, during search, or the main Event store outside of search).

Replace calls to Event.* with references to this in the slot and
schedule functions for slots which may be called outside of search.
(Functions such as Schedule.Validate() should only be called during
search will assume that sched.store is valid.)

Refactor MakeSchedule* a bit: Rather than creating a template schedule
in MakeSchedule(), havind MakeScheduleHeuristic modify it directly and
the others copy it, create a generic function to create a template
from ss.CurrentSchedule and ss.LockedSlots.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>