Make scheduling asynchronous
Remove Schedule.IsStale, and replace it with Event.ScheduleState.
Describe the state transitions and make a set of methods for various
actors to notify of state transition updates.

Refactor the potentially asynchronous part of MakeSchedule into
MakeScheduleAsync.  Add an 'async' argument to MakeSchedule; if this
is true:
- Call MakeScheduleAsync() as a goroutine
- MakeScheduleAsync() must grab the lock before updating the schedule.

Store the state of the scheduler at the time the schedule started; if
it fails for some reason, and nothing has changed it, put it back the
way it was (whether clean or dirty).

Also have MakeSchedule() refuse to start a schedule again if
ScheduleState is InProgress.

For a quick-and-dirty way to get the correct badge, simply set one of
"IsStale", "IsCurrent", or "IsInProgress" when running the template
for the admin console.

While here:
- Make "random" the default search algorithm, "60s" the default
duration
- Say what search algorithm was used in the log message
- Mark the state as modified when changing a discussion's
PossibleSlots and when deleting a discussion.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>