Replace ScheduleState state machine with booleans
The state machine implemenation was broken: if the state was modified
while a search was running, another search could be started.

Instead of having a state machine with enumerated transitions, have
separate booleans for each important bit of state:
 - Are there changes since the last time the schedule was
 updated
 - Are there changes since the schedule "snapshot"
 - Is the scheduler currently running

For simplicity, still only report 3 states on the admin console.

For robustness, clean up "is running" on next load rather than on
shutdown (so that we can recover from power failures / crashes).

Signed-off-by: George Dunlap <george.dunlap@citrix.com>