Implement Active / Non-active mode
Allow the website to run in a "deactivated mode", which allows admins
to do things behind the scenes, but will present others with a special
"check back later" screen.

Add 'Active' as a top-level event property.  When false, don't show
normal toolbar except to logged-in admins; Only show 'Login' button
(for admins to log in).

Add button and code to console screen to change this.

Also add a status indicator on the toolbar, visible only to admins.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>