Fix schedule initialization bug
The proximate bug was that Timetable.Init() was setting all slots at
time "3:25" as a break, rather than all slots at time "3:35".  But the
deeper bug was that we were still using a separate "option" for the
number of slots, rather than counting the slots in the actual
timetable.

The result was that Event.LockedSlots didn't match the number of
slots, causing a crash in Timetable.FillDisplaySlots().

Set the right time for the break, and set store.ScheduleSlots
according to the number of slots actually observed in store.Timetable.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>