Make it possible to disable the schedule
Change template flag from "IsActive" to "IsWebsiteActive", and the
value from 'active' to 'websiteActive'.

Add a new Event flag, "ScheduleActive".  Pass this to templates with
"IsScheduleActive".  If it's not active, hide the 'Schedule' tab, and
return 404 for the "/schedule" url.

Change the "/admin/setactive" form to "/admin/setstatus".  Change the
'name' to 'status' so that we can query all the checkboxes at the same
time.  Add a second checkbox to the form, with the value
"scheduleActive".

Modify "setstatus" handler to correctly handle changes and flashes.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>