discussions: Introduce non-public discussions
Add 'IsPublic' flag to Discussion and DiscussionDisplay.

If 'IsPublic' is false, don't include it in any lists unless the
current user is Admin, or the owner of the discussion.

Set 'IsPublic' to false by default, and on any edits.

For a user, display the public / private status of discussions owned
by them.

For an admin, display public / private status of all discussions, and
also a toggle switch for the public / non-public status.

In order to effect this, pass CurrentUser in to template/item-short.

Add a 'setpublic' action to discussion UIDs and handle it.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>