Add option to not require verification for account registration
Add RequireVerification; allow it to be enabled and disabled on the
console page.

If false, don't display the Verification box on account creation, and
allow accounts to be created which fail the verification check.  If
true, display the Verification box and return an error on the failed
check.

Either way, if the verification check passes, set the account to be
verified.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>