Allow users to enter a verification code.
Add a 'verify' action to user uid post handler.

Add an explanation and a form to the 'user' view allowing the user to
enter the code and posting it to the above handler.

Add a yellow badge next to a user's name when they are unverified.

Also move the User "Edit" button into one of the <div> sections, so
that the layout looks less strange.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>