middleware: Rework routing to avoid login prompt for non-existent URLs
The initial "middleware" (taken from a book on writing web apps in Go)
had a simple list of routers that it would use one-by-one; until
something was handled; it "protected" login-only features by having a
check-and-redirect handler before the router containing those
features.

This sort of worked, except that the result was that non-present URLs
*also* got redirected to login.

Re-implement the middleware with the specific architecture of this
site:
- Public URLs that anybody can see
- "Requires login" URLs that will redirect if not logged in
- "Admin-only" URLs that will return 404 unless you're logged in as an
admin.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>