Add robots.txt
So that search engines know how to handle us (and don't get randomly
redirected to the logon page).

Exclude /discussion/new because robots will just be directed to the
logon page; exclude user list and view because it shouldn't show
anything (and will thus just be a waste of time & bandwidth).

Signed-off-by: George Dunlap <george.dunlap@citrix.com>