Update main website information
First, fix an issue where the context information wasn't being passed
down to template, so logged-in users were seeing info meant for
non-logged-in users.

For non-logged-in users, don't mention the verification code unless
RequreVerification is sit.

For logged-in users:
- "Comment out" previous year's info, but keep for reference
- Copy over info on sessions and profiles
- If verification is not required, explain the unverified / non-public
system
- If the schedule is active, display a more generalized comment about
the schedule; if not, say that the room dates and times have not yet
been determined.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>