Fix error reporting on discussion page
The public/non-public works primarily by having disc.GetDisplay()
return nil when there was no current viewer, or the current viewer
isn't the owner (or admin).

Unfortunately, this broke error reporting on the discussion page, as
disc.GetDisplay() was being called to fill in existing form values in
the template on re-display, and was being passed 'nil'.

Pass back the owner, which will make everything work again.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>