Don't show edit / delete 'GET' pages unless the corresponding 'POST' will work
First, redirect to login when deleting a user as well as editing one.

For safety, add helper functions to check for non-null as well as the
question raised.

Don't display 'discussion edit', 'discussion delete confirmation', or
'user edit' pages if the user may not edit the discussion or user;
don't display the user delete confirmation' page unless the user is an
admin.

While here, make MayEditUser take a user struct rather than a userid,
and check for the validity of the userid before checking if the
current user may edit it.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>