Make "approved" discussion content visible while edits are awaiting moderation
Add "ApprovedTitle" and "ApprovedDescription" elements to the
Discussion struct.

In GetDisplay, if ApprovedTitle is non-"", return it and
ApprovedDescription rather than nil.

On setpublic, copy the title and description into ApprovedTitle and
ApprovedDescription when making public; clear them when making
private.

Add a 'Moderated' badge for users who can't see moderated changes, as
well as a warning at the top of the full-discussion page.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>