Revise registration form
Get rid of moderation warning; a warning will be presented when
proposals are submitted.  State simply that the username is made
public, and move the e-mail address closer to the username.

This should prompt most people not to put their email address in the
username.  The warning can do the rest.

While here, improve logging when new users are rejected due to email
address-like usernames.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>