session: Add tests
Make a `GetCookier` interface which just has the Cookie() method, to
simplify testing.

Change the appropriate functions to accept the above method.

Add a Close() method to SessionStore, and CloseSessionStore() to call
Close() on the "global" session store.  Rename NewSessionStore() to
OpenSessionStore().

Write a test which:
 - Generates 10 random users
 - Randomly performs sessions operations on them keeping track of
   expected state, and checking to see that the expected things
   happen.

Add methods to the structure such that they match the interfaces
required for the functions.

Make a bunch of sessions, close the file, and open it again, making
sure the sessions are in the same state.

One thing this is missing is cookie expiry; that's something to do a
later time.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>