discussions: Move data creation / everything into a sub-package
Move core data-related functionality into a sub-package so that we can
test it separately.

Move "display" functionality into a new file, display.go, in the main
package.

A couple of hacks to get things working for now:
 - main.go: var Event = &disc.Event
 - Don't store the server address
 - No asyncronous schedule creation, so we don't need to deal with the
 lock (yet)

And a couple of bodges:
 - Add methods to access "cached" links in non-exported fields

In anticipation of having sub-package tests, remove some testing
functionality from command-line.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>