serve: Move many flags out of Event and into server keys
TestMode, Active, ScheduleActive, VerificationCodeSent, and
RequireVerification are primarily about the webserver itself, not the
core event functionality.  Remove these booleans from the Event
structure entirely, moving them into keys.

To make using these convenient, add some new functionality to the
keyvalue package:

Exchange, to atomically set a new value, and return the old value (if
any.  Make this robust by creating internal get / set functions that
can be used with transactions.

SetBool, GetBoolDef, and ExchangeBool, for bool-typed keys.
(GetBoolDef ignores errors and returns 'false' unless it finds a key
that says "true".)

In HandleAdminAction(), refactor setting various flags in a cumulative
manner.

The verification code is also really a part of the website, so move it
into the "serve" package as well, setting the default in main() if
it's not present (for now (famous last words)).  Remove vcode checking
in event.NewUser(); instead accept a boolean saying whether the new
user should be pre-verified.  This avoids looking up the verification
code in situations where we know we want the user to be verified as well.

For now, just comment out the TestMode flag in event.TestPopulate();
we'll sort that out when we switch to proper golang-style testing.

While here:
- Report an error if updating testmode fails
- Get rid of redundatn Event.Users.Save() in HandleUserCreate()

Signed-off-by: George Dunlap <george.dunlap@citrix.com>