event: Move all keyvalues into main serve package
The last four keyvalues in 'event' are search parameters.  Move these
into the server code and pass them in as an options struct.

For now, keep a global "opt" variable so we don't need all kinds of
up-pointers.

Make a single function, "MakeSchedule", in the serve code, which will
collect the config information and put it into the struct.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>