event: Move UserProfile up into main 'user' struct
...so that we can use sqlx.Get().

Modify parseProfile() to take a pointer to a user struct and fill in
the "profile" values, rather than returning a "profile" type.

Modify UpdateUser to take a full version of the next user.  Make the
full copy in HandleUidPost edit user code, and have parseProfile()
update the profile as requested.

Modify HandleUserCreate to make the user struct and fill it in
earlier, rather than casting it when making the NewUser call.

To avoid having to reach all the way into the templates for now, make
a scheduler-level UserProfile to pass out to templates.  In a couple
of places, do some hacks to avoid having to restructure the templates
at the current time.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>