event: Don't bother trying to make database creation concurrency-safe
db.Beginx() will sometimes fail with schema busy; but when retrying
that transaction, it still sometimes fails with:

Opening database: db.Beginx: cannot start a transaction within a transaction

This is after modifying go-sqlite3 to always call "ROLLBACK" on any
database begin failure.

This is not something I want to debug now. Let's test transactions and
rollbacks on a transaction we expect might actually race in practice.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>