website: Fix User Edit
A previous refactoring made a custon `UserProfle` struct to be part of
the UserDisplay struct, but neglected to include a Description field;
presumably because there was already a "rendered" Description field
one level up. Unfortunately, this was wrong: the user edit page needs
an unrendered Discription to actually edit.

Add this field to the UserProfile struct and fill it in with a raw value.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>