event: Get rid of location and schedule-related code
We're going to need to rewrite most of it anyway once we switch to
database.

Remove all data structrues from Event.  Remove storeFilename from
EventOptions, and any code handling the old json-formatted code.

Remove ResetData(), as it should be obsoleted by golang testing.

Refactor EventStore.Load() so that all the admin account creation /
password modification code happens in the same place, and is tied to
the existence of the admin account in the database, not to the
presence of the json datafile.

Delete code where it makes sense; put in stubs where it makes sense.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>