event/test: Add discussion creation and interest to the transaction tests
Have each user create 5 discussions.  Have them then get the
discussion by ID again and check its integrity. Finally have each user
get all current discussions, and set an interest in 20 of them (20% of
those being 0), ignoring errors due to since-deleted discussions.

This resulted in the following transaction-related fixes:

* Add a transaction retry loop around DiscussionFindById, UserFind, and
UserFindbyUsername.

* We need to tell when User.SetInterest() failed "legitimately"; i.e.,
the discussion we tried to access has since been deleted.

Add a new error, ErrUserOrDiscussionNotFound, to be returned when a
transaction failed with a Foreign Key constraint violation, and we're
not sure which of the two caused the failure.

Add a new error classifier, isErrorForeignKey(), to return true if the
error returned is a foreign key violation constraint.

Have SetError return ErrUserOrDiscussionNotFound if we get a foreign
key error when setting

* DiscussionIterate is more tricky: apparently the transaction often
(usually? always?) fails when reading rows, rather than from the
initial query.  This makes it tricky, because we definitely don't want
to retry the transaction after we've made the callback for any rows.

Set up a loop in DiscussionIterate where we count the number of
callbacks made successfully.  Check rows.Err(); only retry the whole
transaction we haven't processed any rows yet.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>