event/test: Add more Discussion transaction tests
Add tests for DiscussionUpdate, DiscussionSetPublic, and
Discussion.GetMaxScore().

Also add a test for UserFindByUsername.

Refactor DiscussionSetPublic to use the same code with just a
different query; make this code restartable.

Also fix an error in the goroutine harvesting code: having shouldExit
on the LHS of the || means that once it becomes false, it skips any
subsequent channel reads and exits immediately, causing panics writing
on closed channels.  Move the channel read to the left so that it
always happens.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>