event/db: Add location
Add a 'location' field to event_users, constraining it not to be NULL,
and specifying that it should be parseable by time.LoadLocation().

Also specify that slots be the output of time.MarshalText(), which
will ensure that the slots have timezone location embedded in them.
(No other change required at the moment since nothing accesses these
yet.)

Define TZLocation, a wrapper around time.Location that has SQL
marshalling / unmarshalling code.

Add a Location field of type TZLocation to the User struct.

Add a default location string to EventOpts and require its presence;
store that in EventStore.

When creating a user, if the Location pointer is nil, use the default
location.

Add code to insert / update  location when appropriate.

Add the default location as a keyvar in main.

Test creating users either with no location, or with the default
location.  Try both leaving the value alone or changing it.

(While here fix an error message in compareUsers.)

Signed-off-by: George Dunlap <george.dunlap@citrix.com>