ui: Add web code to add locations
Add a new "locations" tab to the admin console, which lists current
locations in a form suitable to update, as well as an empty slot to
add new locations.

Add code to render the new page, as well as code to handle new POST
values.

Note that from a UI perspective this isn't great: Our error reporting
stinks and we don't retain content on submission failure.  Just get it
working first and improve it later.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>