event/timetable: Add full timetime
Update sqlx version to get a bug fix.

Fix a typo in the schema for event_schedule, which causes it to
reference event_slots:locationid rather than
event_locations:locationid.

Sync previous schema updates to event_slots (changing slotid to text,
adding slotidx, and changing the uniqueness constraint) to the
database creation and upgrade routines.

On upgrade, just drop & regenerate both tables.

Create a `txLoop` helper function, to loop over a function as long as
it returns a shouldRetry() error.  Port a couple of functions in
location.go over to this so we can test it.

Clone time.Time into our timedb code, so that we can automatically
marshal and unmarshall it.  Change TimetableSlot.Time to this type.

Rename TimetableDiscussion.ID to DiscussionID for clarity.

Implement the day / slot parts of GetTimetable().

Add testing.

Signed-off-by: George Dunlap <george.dunalp@citrix.com>