Format time for a specific location
Present a formatted time in a specific location on the schedule page.

Rather than marshal the whole structure over just to be able to format
the time string, add a TimeDisplay element to the event.TimetableSlot
structure.

Modify GetTimetable to take a format and location, and have it fill in
TimeDisplay appropriately.

Rather than trying to pass around the location, make DefaultLocation a
global variable, and set that after handling the database /
commandline / default logic.  Also look up the default location and
use that as a TZ.

Use the default location, and format the string to contain the time
first, and then the date; time because that's the most important, date
because in the extremes people may end up crossing over timezones.

While here, fix the editTimetable command to include the default
location when the current timetable is non-empty.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>