serve: Expose possible scheduling slots again
Fix a couple of places where we should be using DiscussionFull rather
than Discussion.

Add SlotsSetTimeDisplay to format slot times for DiscussionDisplay.
If Discussion.Time is non-zero, set TimeDisplay.

Use new DisplaySlot elements for checklist:
- SlotID rather than Index
- TimeDisplay rather than Label

Fix templates to use new DiscussionDisplay elements:
- OwnerInfo rather than Owner
- Owner rather than Owner.UserID
- LocationName
- TimeDisplay rather than Time

The way checkbox forms work is that they only send the checked boxes;
so refactor DiscussionSetPossibleSlots to take a list of []SlotIDs.
These will be valid slots; all other slots will not be.  Refactor
possibleslot_test.go with this new way of doing things.

Replace FormCheckToBool with FormcheckToSlotID.  After updating the
discussion, set possible slots as well.

Add DiscussionGetDisplayRetry(), which is used for validation errors
when creating a new discussion or editing existing discussions. When
editing existing discussions is done by an admin, it needs OwnerInfo,
AllUsers, and PossibleSlots.  Modify HandleDiscussionCreate to mock up
a DiscussionFull so that we can use the same function for  both it and edit.

While here:

- Refactor another place to use "discussion/slots-form" rather than
  open-coding it

- Fix a place where we were still using .Profile

Signed-off-by: George Dunalp <george.dunlap@citrix.com>