event: Add heuristic schedule
Make searchDiscussion.PossibleSlots a map instead of a slot, to make
it easier to query whether a given slot is possible for a given
discussion.

Fix scheduleSet to handle schedules with empty slots.

Add a helper to create an empty schedule for the scheduler to fill in.

Add helper functions to calculate the effect of adding a discussion to
a specific slot.

Add tests for heuristic scheduler.

While here, add the attendees and interest to the GetTimetable query.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>