event: Use location url rather than location description
locationdescription -> locationurl
LocationDescription -> LocationURL
locDisc -> locURL

In TimetableDiscussion, replace generic LocationInfo with LocationName
and LocationString, so that these can be filled in with the main
discussion info query.

In the discussion info query, merge with event_locations at the top,
pulling out locatinoname and locationurl, and passing them up through
the queries to the top level.

Make a link to the location; add a "Room" label to distinguish it from
the title and make clear what the link is for.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>