serve: Make it possible to set the timezone
Add a sub-package which will take a "cleaned" list of timezones, and
check to make sure each one works before returning it in a list.

Add a selection form to the top of schedule.  Cause this to use 'GET'
on the same URL for simplicity for now.

Have the scheduler handler look for 'location' in the URL form
values.  If it finds one, look up the location; and if this succeeds,
use this as the timezone; otherwise, revert to the default.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>