handlers: Wire up the ability to set default location
Add User.GetLocationTZ() to return the location stored in the DB.
Test this in the transaction tests.

Replace the cookie-reading code in handle_schedule.go with code which
checks for a logged-in user's default location.

Add timezone drop-downs to the user creation and edit pages.  Arrange
to pass default tz and timezone lists to these forms.

Parse the location in parseProfile (used both by new and edit).

Display a user's default timezone to anyone capable of editing that
user (i.e., themselves or admin).

While here, fix the address of the email editing.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>