ui: Add timezones to discussions and slots
Add a TZ offset to slot displays (for both locked slots and possible
scheduled slots), in the default timezone, to remind the administrator
what timezone to think in.

Add a TZ offset as well as the named timezone when displaying a
discussion by itself.  If there's a user logged in, display it in
their default timezone.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>