serve: Close databases when shutting down
Since we're using locking, it's probably safe to just exit; but it
does leave the WAL log around for the eventdb.  Again, keyvalue and
sessions aren't using WAL at the moment, but better to make things
clean.

A more robust approach would be to allow the places which open the
files to register a hook to shut things down; something to do for the
future.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>