event: Place discussions even if there isn't a particularly good place
In situations where placing a session never increases the score (which
could mainly happen when placing fixed positions made by an admin),
the algorithm mistakenly concluded there was nowhere to place it.

Fix this by adding the number of empty locations to the score.  This
will guarantee that placing a topic will always be better than not
placing it, and will additionally favor a more evened-out schedule
when all other factors are equal.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>