event: Don't limit number of discussions administrators can make
The most common case will be that the user is not an admin, so we need
to do the check; therefore count discussions first, then only if
they're exceeded, check to see if the user is an admin.

Add some tests to make sure it works (and stays working).

While here, get rid of some spurious formatting from errOrRetry.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>